// packages/spotify-effect/src/model/SpotifyOptions.ts
export type MarketOptions = {
  market?: string;
};

export type GetAlbumTracksOptions = {
  limit?: number;
  offset?: number;
  market?: string;
};